% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoApp.R
\name{demoApp}
\alias{demoApp}
\title{demoApp}
\usage{
demoApp(topic)
}
\arguments{
\item{topic}{character: example app to run}
}
\value{
nothing
}
\description{
Runs a Shiny demo app or shows an overview of all Shiny demo apps related to \code{topic}. 
If more than one app relates to \code{topic} a list of apps is shown otherwise the app is run.
}
\examples{
demoApp() 
demoApp('gettext') 
\dontrun{
  demoApp('gettext::app2en') 
  demoApp('app2de') 
}
}
