% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsub.R
\name{dsub}
\alias{dsub}
\title{dsub}
\usage{
dsub(txt, ...)
}
\arguments{
\item{txt}{character: text to replace in}

\item{...}{named parameter or list with pattern and replacement}
}
\value{
replaced text
}
\description{
Performs dollar substitution in a text.
}
\examples{
dsub('$a EUR', a=3)
dsub('$a EUR', list(a=3))
}
