% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme.md.R
\name{readme.md}
\alias{readme.md}
\title{readme.md}
\usage{
readme.md(path = ".", env = env_package(path))
}
\arguments{
\item{path}{Either specify a \code{path} to the root directory of
a package, an R \code{file}, or a character vector \code{text}.}

\item{env}{An environment environment containing the result of evaluating
the input code. The defaults will do this for you in a test environment:
for real code you'll need to generate the environment yourself.

You can also set to \code{NULL} if you only want to get the tokenized code
blocks only. This suppresses evaluation of \verb{@eval} tags, and will not
find the code object associated with each block.}
}
\value{
strings with markdown code
}
\description{
Creates a lines for a README.md for Github based on roxygen2 comments and uses \code{\link[roxygen2]{parse_package}}.
}
