% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoApp.R
\name{demoApp}
\alias{demoApp}
\title{demoApp}
\usage{
demoApp(topic = NULL, package = "gettext")
}
\arguments{
\item{topic}{character: example(s) to show or \code{NULL}}

\item{package}{character: package to use (default: \code{"gettext"})}
}
\value{
nothing
}
\description{
Runs one Shiny demo app of a package. With \code{demoApp(package="gettext")} 
you get a vector of topics back of the package \code{gettext}.
}
\examples{
demoApp(package='gettext') 
\dontrun{
  demoApp('app2en', 'gettext') 
  demoApp('app2de', 'gettext') 
}
}
